Refactor add external account modalize
