+++ src/ui/screens/AccountsScreen.tsx
@@ -6,19 +6,14 @@ import {Account as AccountType, useAccounts} from 'context/index';
 import {useTheme, Divider, IconButton, List, FAB, Caption, Menu, Subheading, Icon, useBottomSheet} from '@ui/library';
 import SafeView, {noTopEdges} from '@ui/components/SafeView';
 import {CompleteNavigatorParamList} from '@ui/navigation/navigation';
-import {
-  accountsScreen,
-  addAccountScreen,
-  importAccountScreen,
-  mnemonicScreen,
-  myAccountScreen,
-} from '@ui/navigation/routeKeys';
+import {accountsScreen, importAccountScreen, mnemonicScreen, myAccountScreen} from '@ui/navigation/routeKeys';
 import globalStyles, {standardPadding} from '@ui/styles';
 import {EmptyView} from '@ui/components/EmptyView';
 import {useAccount} from 'src/api/hooks/useAccount';
 import {Account} from '@ui/components/Account/Account';
 import {Padder} from '@ui/components/Padder';
 import {AccountsGuide} from '@ui/components/Account/AccountsGuide';
+import {AddExternalAccount} from '@ui/components/Account/AddExternalAccount';
 
 type Props = {
   navigation: NavigationProp<CompleteNavigatorParamList, typeof accountsScreen>;
@@ -28,6 +23,11 @@ type SortBy = 'name' | 'favorites';
 
 export function AccountsScreen({navigation}: Props) {
   const {openBottomSheet, BottomSheet} = useBottomSheet();
+  const {
+    openBottomSheet: openAddAccount,
+    closeBottomSheet: closeAddAccount,
+    BottomSheet: AddAccount,
+  } = useBottomSheet();
   const {networkAccounts, toggleFavorite} = useAccounts();
 
   const [sortBy, setSortBy] = React.useState<SortBy>('name');
@@ -99,11 +99,15 @@ export function AccountsScreen({navigation}: Props) {
         ItemSeparatorComponent={Divider}
         ListEmptyComponent={EmptyView}
       />
-      <Buttons navigation={navigation} />
+      <Buttons navigation={navigation} onAddAccount={openAddAccount} />
 
       <BottomSheet>
         <AccountsGuide />
       </BottomSheet>
+
+      <AddAccount>
+        <AddExternalAccount onClose={closeAddAccount} />
+      </AddAccount>
     </SafeView>
   );
 }
@@ -172,7 +176,12 @@ function AccountItem({
   );
 }
 
-const Buttons = ({navigation}: {navigation: Props['navigation']}) => {
+type FabProps = {
+  navigation: Props['navigation'];
+  onAddAccount: () => void;
+};
+
+const Buttons = ({navigation, onAddAccount}: FabProps) => {
   const [state, setState] = React.useState({open: false});
   const onStateChange = ({open}: {open: boolean}) => setState({open});
   const {open} = state;
@@ -191,7 +200,7 @@ const Buttons = ({navigation}: {navigation: Props['navigation']}) => {
         {
           icon: 'plus',
           label: 'Add External Account',
-          onPress: () => navigation.navigate(addAccountScreen),
+          onPress: onAddAccount,
         },
         {
           icon: 'key-plus',
